refactor(plugin-rows): clean up components
